requirements.txt
click==8.1.7
fastapi==0.114.0
httpx==0.27.2
numpy==2.1.1
openai==1.44.0
pinecone==5.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
uvicorn==0.30.6
